+++ src/components/AccountActivityDataCellContractCreateTxEvent.vue
@@ -1,0 +1,20 @@
+<template>
+  <value-hash-ellipsed
+    :hash="activity.payload.tx.contractId"
+    :link-to="`/contracts/${activity.payload.tx.contractId}`"/>
+</template>
+
+<script setup>
+import ValueHashEllipsed from '@/components/ValueHashEllipsed'
+
+defineProps({
+  activity: {
+    required: true,
+    type: Object,
+  },
+  accountDetails: {
+    type: Object,
+    default: null,
+  },
+})
+</script>