+++ src/components/TokenEventsTableCondensed.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <table
-      v-for="event in tokenEvents.data"
+      v-for="(event, index) in tokenEvents.data"
       :key="event.hash"
       class="token-events-table-condensed__table">
       <tbody>
@@ -54,14 +54,37 @@
         </tr>
 
         <tr class="token-events-table-condensed__row">
-          <th class="token-events-table-condensed__header">
+          <th
+            :class="[
+              'token-events-table-condensed__header',
+              {'token-events-table-condensed__header--expanded': isExpanded.includes(index)}]">
             Data
           </th>
-          <td class="token-events-table-condensed__data">
+          <td
+            v-if="event.isDecoded"
+            class="token-events-table-condensed__data">
             <token-events-data-cell
               :name="event.name"
               :data="event.data"
               :args="event.args"/>
+          </td>
+          <td
+            v-else
+            :class="[
+              'token-events-table-condensed__data',
+              {'token-events-table-condensed__data--expanded': isExpanded.includes(index)}]">
+            <expand-button
+              :is-expanded="isExpanded.includes(index)"
+              @click="toggle(index)">
+              {{ isExpanded.includes(index) ? 'Hide arguments' : 'See arguemnts' }}
+            </expand-button>
+          </td>
+        </tr>
+        <tr
+          v-if="isExpanded.includes(index)"
+          class="token-events-table-condensed__row">
+          <td colspan="5">
+            <event-data-panel :args="event.args"/>
           </td>
         </tr>
       </tbody>
@@ -71,15 +94,31 @@
 <script setup>
 import ValueHashEllipsed from '@/components/ValueHashEllipsed'
 import DatetimeLabel from '@/components/DatetimeLabel'
-import TokenEventsDataCell from '@/components/TokenEventsDataCell.vue'
+import TokenEventsDataCell from '@/components/TokenEventsDataCell'
 import { tokensHints } from '@/utils/hints/tokensHints'
+import ExpandButton from '@/components/ExpandButton'
 
-defineProps({
+const props = defineProps({
   tokenEvents: {
     type: Object,
     required: true,
   },
 })
+
+const isExpanded = ref([])
+
+watch(() => props.tokenEvents, () => {
+  isExpanded.value = []
+})
+
+function toggle(id) {
+  const index = isExpanded.value.indexOf(id)
+  if (index > -1) {
+    isExpanded.value.splice(index, 1)
+  } else {
+    isExpanded.value.push(id)
+  }
+}
 </script>
 
 <style scoped>
@@ -91,6 +130,10 @@ defineProps({
 
   &__header {
     border-bottom: 1px solid var(--color-midnight-25);
+
+    &--expanded {
+      border-bottom: 0;
+    }
   }
 
   &__row:last-of-type &__header {
@@ -99,6 +142,10 @@ defineProps({
 
   &__data {
     text-align: right;
+
+    &--expanded {
+      border-bottom: 0;
+    }
   }
 }
 </style>