+++ src/stores/accountDetails.js
@@ -2,7 +2,7 @@ import { defineStore, storeToRefs } from 'pinia'
 import { useRuntimeConfig } from 'nuxt/app'
 import useAxios from '@/composables/useAxios'
 import { useMarketStatsStore } from '@/stores/marketStats'
-import { adaptAccountNames, adaptAccountTokens, adaptTransactions } from '@/utils/adapters'
+import { adaptAccountActivities, adaptAccountNames, adaptAccountTokens, adaptTransactions } from '@/utils/adapters'
 import { formatAettosToAe } from '@/utils/format'
 import { useDexStore } from '@/stores/dex'
 
@@ -16,7 +16,6 @@ export const useAccountStore = defineStore('account', () => {
   const { fetchPrice } = useDexStore()
 
   const rawAccountDetails = ref(null)
-  const accountTransactionsCount = ref(null)
   const totalAccountTransactionsCount = ref(null)
   const accountNamesCount = ref(null)
   const selectedKeyblock = ref(null)
@@ -24,6 +23,7 @@ export const useAccountStore = defineStore('account', () => {
   const selectedKeyblockMicroblocks = ref(null)
   const selectedMicroblockTransactions = ref(null)
   const rawAccountNames = ref(null)
+  const rawAccountActivities = ref(null)
   const rawAccountTokens = ref(null)
   const rawAccountTransactions = ref(null)
   const tokenPrices = ref({})
@@ -33,11 +33,15 @@ export const useAccountStore = defineStore('account', () => {
       ? {
         ...rawAccountDetails.value,
         balance: formatAettosToAe(rawAccountDetails.value.balance),
-        transactionsCount: accountTransactionsCount.value,
         totalTransactionsCount: totalAccountTransactionsCount.value,
         namesCount: accountNamesCount.value,
         isGeneralized: rawAccountDetails.value.kind === 'generalized',
       }
+      : null,
+  )
+  const accountActivities = computed(() =>
+    rawAccountActivities.value
+      ? adaptAccountActivities(rawAccountActivities.value)
       : null,
   )
   const accountTransactions = computed(() =>
@@ -67,6 +71,7 @@ export const useAccountStore = defineStore('account', () => {
         fetchTotalAccountTransactionsCount(accountId),
         fetchAccountNames({ accountId, limit }),
         fetchAccountNamesCount(accountId),
+        fetchAccountActivities({ accountId, limit }),
       ]),
     ])
     return true
@@ -81,17 +86,6 @@ export const useAccountStore = defineStore('account', () => {
         rawAccountDetails.value = { id: accountId, notExistent: true }
       }
     }
-  }
-
-  async function fetchAccountTransactionsCount(accountId, txType = null) {
-    accountTransactionsCount.value = null
-
-    const params = txType ? `/${accountId}?type=${txType}` : `?id=${accountId}`
-    const txCountUrl = new URL(`${MIDDLEWARE_URL}/v2/txs/count${params}`)
-
-    const { data } = await axios.get(txCountUrl)
-
-    accountTransactionsCount.value = data
   }
 
   async function fetchTotalAccountTransactionsCount(accountId) {
@@ -137,6 +131,13 @@ export const useAccountStore = defineStore('account', () => {
     accountNamesCount.value = data.data.length
   }
 
+  async function fetchAccountActivities({ accountId, limit, queryParameters } = {}) {
+    rawAccountActivities.value = null
+    const defaultParameters = `/v2/accounts/${accountId}/activities?limit=${limit ?? 10}`
+    const { data } = await axios.get(`${MIDDLEWARE_URL}${queryParameters || defaultParameters}`)
+    rawAccountActivities.value = data
+  }
+
   async function fetchAccountTransactions({ accountId, type, limit, queryParameters } = {}) {
     rawAccountTransactions.value = null
 
@@ -151,7 +152,7 @@ export const useAccountStore = defineStore('account', () => {
     transactionsUrl.searchParams.append('limit', limit ?? 10)
 
     if (accountId) {
-      transactionsUrl.searchParams.append('account', accountId)
+      transactionsUrl.searchParams.append('sender_id', accountId)
     }
 
     if (type) {
@@ -164,7 +165,6 @@ export const useAccountStore = defineStore('account', () => {
 
   return {
     rawAccountDetails,
-    accountTransactionsCount,
     totalAccountTransactionsCount,
     accountNamesCount,
     selectedKeyblock,
@@ -172,17 +172,19 @@ export const useAccountStore = defineStore('account', () => {
     selectedKeyblockMicroblocks,
     selectedMicroblockTransactions,
     rawAccountNames,
+    rawAccountActivities,
     rawAccountTransactions,
     rawAccountTokens,
     accountDetails,
+    accountActivities,
     accountTransactions,
     accountNames,
     accountTokens,
     tokenPrices,
     fetchAccount,
+    fetchAccountActivities,
     fetchTotalAccountTransactionsCount,
     fetchAccountTransactions,
-    fetchAccountTransactionsCount,
     fetchAccountNames,
     fetchAccountTokens,
   }