+++ src/utils/constants.js
@@ -73,3 +73,19 @@ export const TOKEN_SUPPLY_ACI = [
     },
   },
 ]
+export const SWAP_CONTRACT_CALLS = [
+  'swap_exact_tokens_for_tokens',
+  'swap_tokens_for_exact_tokens',
+  'swap_exact_ae_for_tokens',
+  'swap_tokens_for_exact_ae',
+  'swap_exact_tokens_for_ae',
+  'swap_ae_for_exact_tokens',
+]
+export const ADD_LIQUIDITY_CONTRACT_CALLS = [
+  'add_liquidity',
+  'add_liquidity_ae',
+]
+export const REMOVE_LIQUIDITY_CONTRACT_CALLS = [
+  'remove_liquidity',
+  'remove_liquidity_ae',
+]